requirements.txt
pytz
tzlocal
datetime
requests
climage
python-dateutil
readchar
colorama
ruamel.yaml
rich
websockets